refactor trace timeline to grid
