+++ packages/cli/src/playground/src/pages/observability/trace-tree.tsx
@@ -1,0 +1,39 @@
+import { cn } from '@/lib/utils';
+import { TraceTreeSpan } from './trace-tree-span';
+
+type TraceTreeProps = {
+  spans?: any[];
+  onSpanClick: (spanId: string) => void;
+  selectedSpanId?: string;
+  overallLatency: number;
+  overallStartTime: string;
+};
+
+export function TraceTree({
+  spans = [],
+  onSpanClick,
+  selectedSpanId,
+  overallLatency,
+  overallStartTime,
+}: TraceTreeProps) {
+  return (
+    <div
+      className={cn(
+        'overflow-y-auto pr-[1.5rem] grid items-start content-start gap-y-[2px] xl:py-[1.5rem] ',
+        'xl:grid-cols-[3fr_2fr] xl:gap-x-[1rem]',
+        //    'border border-red-500',
+      )}
+    >
+      {spans?.map((span: any) => (
+        <TraceTreeSpan
+          key={span.id}
+          span={span}
+          onSpanClick={onSpanClick}
+          selectedSpanId={selectedSpanId}
+          overallLatency={overallLatency}
+          overallStartTime={overallStartTime}
+        />
+      ))}
+    </div>
+  );
+}
